refactor: 타입 변경 및 mock 데이터 변경
