+++ src/app/(navigation)/home/_hooks/queries/useHotAuctionRecommends.ts
@@ -1,11 +1,10 @@
 import { useQuery } from "@tanstack/react-query";
 
-import { getHotAuctionRecommends } from "@/app/api/getAuctionRecommends";
-
-import { Root } from "../../_component/DetailAuctionFeed";
+import { getHotAuctionRecommends } from "@/app/(navigation)/home/_api/getAuctionRecommends";
+import { RecommendAuctionsResponse } from "@/utils/types/auction/recommendAuction";
 
 const useHotAuctionRecommends = () => {
-  const { data, isPending } = useQuery<Root>({
+  const { data, isPending } = useQuery<RecommendAuctionsResponse>({
     queryKey: ["auction"],
     queryFn: getHotAuctionRecommends,
     staleTime: 60 * 1000,
