+++ src/utils/mocks/http.ts
@@ -1,6 +1,7 @@
 import { createMiddleware } from "@mswjs/http-middleware";
-import express from "express";
 import cors from "cors";
+import express from "express";
+
 import { handlers } from "./handlers";
 
 const app = express();
